+++ src/FullHealth/Configuration.cs
@@ -1,0 +1,9 @@
+using BepInEx.Configuration;
+
+namespace FullHealth;
+
+public static class Configuration
+{
+    public static ConfigEntry<bool> Enabled { get; set; }
+    public static ConfigEntry<decimal> Percent { get; set; }
+}