+++ src/FullHealth/Plugin.cs
@@ -1,4 +1,5 @@
-﻿using BepInEx;
+﻿using System;
+using BepInEx;
 using BepInEx.Logging;
 using HarmonyLib;
 
@@ -13,8 +14,13 @@ public class Plugin : BaseUnityPlugin
     private void Awake()
     {
         Logger = base.Logger;
-        Logger.LogInfo($"Plugin {MyPluginInfo.PLUGIN_GUID} is loaded!");
+        Configuration.Enabled = Config.Bind("General", "Enabled", true, "Indicates whether the plugin is enabled");
+        Configuration.Percent = Config.Bind("General", "Percent", 100M, "Percentage of maximum health to which the player will be healed (Min: 0, Max: 100)");
+        Configuration.Percent.Value = Math.Min(100M, Math.Max(0M, Configuration.Percent.Value));
+
         var harmony = new Harmony(MyPluginInfo.PLUGIN_GUID);
-        harmony.PatchAll(typeof(FullHealthPatch));
+        harmony.PatchAll(typeof(FullHealthMultiPlayerPatch));
+        harmony.PatchAll(typeof(FullHealthSinglePlayerPatch));
+        Logger.LogInfo($"Plugin {MyPluginInfo.PLUGIN_GUID} is loaded!");
     }
 }