+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Change Log
 
+## [1.1.0] - 2025-03-30
+
+### Added
+
+- `Enabled` configuration parameter which indicates whether the plugin is enabled
+- `Percent` configuration parameter which set percentage of maximum health to which the player will be healed
+
+### Fixed
+
+- Functionality of plugin for single player and multiplayer
+
 ## [1.0.0] - 2025-03-30
 
 ### Added
 
-The first version of the plugin 1.0.0 has been added
+- The first version of the plugin 1.0.0 has been added